+++ src/main.c
@@ -11,8 +11,8 @@
 #include <string.h>
 #include <sys/stat.h>
 
-#include "../inc/amd_command_buffer_handler.h"
-#include "../inc/amd_command_ring_buffer.h"
+#include "../inc/amd64_command_buffer_handler.h"
+#include "../inc/amd64_command_ring_buffer.h"
 #include "../inc/powerpc_virtual_storage_model.h"
 
 #define __iomem                                                                \
