+++ src/wrangling.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "../inc/amd_command_buffer_handler.h"
-#include "../inc/amd_command_ring_buffer.h"
+#include "../inc/amd64_command_buffer_handler.h"
+#include "../inc/amd64_command_ring_buffer.h"
